#88 feat: DeleteTrashButton 컴포넌트 작성
