+++ src/components/mypage/MyLogContent/index.tsx
@@ -1,6 +1,9 @@
+'use client';
+
 import MyLog from '../Log';
 
 import { DEFAULT_CONTENTS_SIZE } from '@constants/mypage';
+import { useDeleteLog } from '@hooks/api/useDeleteLog';
 import { useInfiniteGetMyLog } from '@hooks/api/useInfiniteMyLog';
 import useObserver from '@hooks/useObserver';
 
@@ -13,6 +16,7 @@ export default function MyLogContent() {
   } = useInfiniteGetMyLog({
     size: DEFAULT_CONTENTS_SIZE,
   });
+  const { mutate: deleteLog } = useDeleteLog();
 
   const onIntersect: IntersectionObserverCallback = ([entry]) => {
     if (entry.isIntersecting) fetchNextPage();
@@ -23,13 +27,23 @@ export default function MyLogContent() {
     threshold: 1,
   });
 
+  const handleDeleteLog = (id: number) => {
+    deleteLog(id);
+  };
+
   return (
-    <div className="mx-[16px]">
+    <div className="mx-[16px] overflow-y-auto h-[calc(100dvh-361px)]">
       {myLog &&
         !myLog[0].data.empty &&
         myLog?.map((page) => {
           return page.data.content.map((item) => {
-            return <MyLog key={item.reviewId} {...item} />;
+            return (
+              <MyLog
+                key={item.reviewId}
+                onClick={() => handleDeleteLog(item.reviewId)}
+                {...item}
+              />
+            );
           });
         })}
       {!isLoading && hasNextPage && <div ref={setTarget} />}
